backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.24.0
psycopg2-binary>=2.9.0
prophet>=1.1.0
scikit-learn>=1.0.0
pandas>=2.0.0
numpy>=1.20.0
requests>=2.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
